pyproject.toml
[tool.poetry]
name = "qadiagbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.5"
google-cloud = "^0.34.0"
google-cloud-storage = "^1.38.0"
dialogflow = "^1.1.0"
google-cloud-dialogflow = "^2.3.0"
vk-api = "^11.9.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
